Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -2,9 +2,9 @@
 
 
 int main() {
-    double a,b;
-    scanf("%d",a);
-    scanf("%d",b);
-    printf("%d", a+b);
+    int a,b;
+    scanf("%i",a);
+    scanf("%i",b);
+    printf("%i", a+b);
     return 0;
 }
